refactor: di module 패키징화 적용
