+++ app/src/main/java/com/preonboarding/sensordashboard/di/coroutine/SensorScopeQualifier.kt
@@ -1,4 +1,4 @@
-package com.preonboarding.sensordashboard.di
+package com.preonboarding.sensordashboard.di.coroutine
 
 import javax.inject.Qualifier
